refactor(selector): break resolver import cycle
SelectorResolverPipeline.frames and .main imported the value
tryInContextInternal through the SelectorResolver barrel, which is
a member of the 5-node Selector SCC. Those two barrel imports were
the only back-edges into the barrel, closing the cycle.

Redirect both to the definition leaf SelectorResolver.try (a clean
out-of-SCC module: playwright + LoginConfig + probe/types only).
Symbol identity is unchanged (the barrel re-exports the same
binding), so behavior is byte-identical and type-only at the edge.

This dissolves the entire 5-node Selector cycle: coupling mass
16 -> 11, cycleCount 2 -> 1. Baseline re-frozen subtractively
(only the Selector cycle removed; the ApiMediator cycle is intact).
