+++ src/Tests/Tools/import-cycles.baseline.json
@@ -1,6 +1,6 @@
 {
   "note": "Frozen dependency cycles (Tarjan SCCs, size ≥ 2). A current cycle passes only when it is a subset of one listed here. Burn down toward [].",
-  "cycleCount": 2,
+  "cycleCount": 1,
   "cycles": [
     [
       "src/Scrapers/Pipeline/Mediator/Api/ApiMediator.builders.ts",
@@ -14,13 +14,6 @@
       "src/Scrapers/Pipeline/Mediator/Api/ApiMediator.types.ts",
       "src/Scrapers/Pipeline/Mediator/Api/ITokenStrategy.ts",
       "src/Scrapers/Pipeline/Mediator/Api/TokenResolverBuilder.ts"
-    ],
-    [
-      "src/Scrapers/Pipeline/Mediator/Selector/SelectorResolver.entries.ts",
-      "src/Scrapers/Pipeline/Mediator/Selector/SelectorResolver.ts",
-      "src/Scrapers/Pipeline/Mediator/Selector/SelectorResolverPipeline.frames.ts",
-      "src/Scrapers/Pipeline/Mediator/Selector/SelectorResolverPipeline.main.ts",
-      "src/Scrapers/Pipeline/Mediator/Selector/SelectorResolverPipeline.ts"
     ]
   ]
 }
